requirements.txt
openai>=1.0.0
pytesseract
Pillow
python-dotenv
fastapi
uvicorn
pydantic
tqdm
aiofiles
httpx
loguru
opencv-python
pdf2image
python-multipart
pytesseract
easyocr
paddlepaddle==3.1.1
paddleocr==3.2.0
Pillow
pandas
openpyxl
matplotlib
